Refactor toast commponents
